Use TransData2D to perform the frame motion

+++ source/blender/blenkernel/BKE_grease_pencil.hh
@@ -234,7 +234,7 @@ class LayerRuntime {
    */
   Vector<LayerMask> masks_;
 
-  Map<int, int> trans_frame_data_;
+  Map<int, float> trans_frame_data_;
   Map<int, GreasePencilFrame> trans_frames_copy_;
 };
